+++ packages/block-library/src/tag-cloud/index.js
@@ -12,17 +12,12 @@ export const name = 'core/tag-cloud';
 
 export const settings = {
 	title: __( 'Tag Cloud' ),
-
 	description: __( 'A cloud of your most used tags.' ),
-
 	icon: 'tag',
-
 	category: 'widgets',
-
 	supports: {
 		html: false,
 		align: true,
 	},
-
 	edit,
 };
