+++ packages/block-library/src/template/edit.js
@@ -1,0 +1,8 @@
+/**
+ * WordPress dependencies
+ */
+import { InnerBlocks } from '@wordpress/block-editor';
+
+export default function TemplateEdit() {
+	return <InnerBlocks />;
+}
