+++ packages/block-library/src/template/index.js
@@ -2,13 +2,14 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { InnerBlocks } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
+import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
 
 const { name } = metadata;
 
@@ -16,22 +17,13 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Reusable Template' ),
-
 	description: __( 'Template block used as a container.' ),
-
 	icon,
-
 	supports: {
 		customClassName: false,
 		html: false,
 		inserter: false,
 	},
-
-	edit() {
-		return <InnerBlocks />;
-	},
-
-	save() {
-		return <InnerBlocks.Content />;
-	},
+	edit,
+	save,
 };
