+++ packages/block-library/src/text-columns/index.js
@@ -1,20 +1,15 @@
 /**
- * External dependencies
- */
-import { get, times } from 'lodash';
-
-/**
  * WordPress dependencies
  */
-import { createBlock } from '@wordpress/blocks';
 import { __ } from '@wordpress/i18n';
-import { RichText } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
 import edit from './edit';
 import metadata from './block.json';
+import save from './save';
+import transforms from './tranforms';
 
 const { name } = metadata;
 
@@ -25,56 +20,15 @@ export const settings = {
 	supports: {
 		inserter: false,
 	},
-
 	title: __( 'Text Columns (deprecated)' ),
-
 	description: __( 'This block is deprecated. Please use the Columns block instead.' ),
-
-	transforms: {
-		to: [
-			{
-				type: 'block',
-				blocks: [ 'core/columns' ],
-				transform: ( { className, columns, content, width } ) => (
-					createBlock(
-						'core/columns',
-						{
-							align: ( 'wide' === width || 'full' === width ) ? width : undefined,
-							className,
-							columns,
-						},
-						content.map( ( { children } ) =>
-							createBlock(
-								'core/column',
-								{},
-								[ createBlock( 'core/paragraph', { content: children } ) ]
-							)
-						)
-					)
-				),
-			},
-		],
-	},
-
+	transforms,
 	getEditWrapperProps( attributes ) {
 		const { width } = attributes;
 		if ( 'wide' === width || 'full' === width ) {
 			return { 'data-align': width };
 		}
 	},
-
 	edit,
-
-	save( { attributes } ) {
-		const { width, content, columns } = attributes;
-		return (
-			<div className={ `align${ width } columns-${ columns }` }>
-				{ times( columns, ( index ) =>
-					<div className="wp-block-column" key={ `column-${ index }` }>
-						<RichText.Content tagName="p" value={ get( content, [ index, 'children' ] ) } />
-					</div>
-				) }
-			</div>
-		);
-	},
+	save,
 };
