+++ packages/block-library/src/verse/index.js
@@ -2,8 +2,6 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { createBlock } from '@wordpress/blocks';
-import { RichText } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
@@ -11,6 +9,8 @@ import { RichText } from '@wordpress/block-editor';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
+import transforms from './tranforms';
 
 const { name } = metadata;
 
@@ -18,46 +18,12 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Verse' ),
-
 	description: __( 'Insert poetry. Use special spacing formats. Or quote song lyrics.' ),
-
 	icon,
-
 	keywords: [ __( 'poetry' ) ],
-
-	transforms: {
-		from: [
-			{
-				type: 'block',
-				blocks: [ 'core/paragraph' ],
-				transform: ( attributes ) =>
-					createBlock( 'core/verse', attributes ),
-			},
-		],
-		to: [
-			{
-				type: 'block',
-				blocks: [ 'core/paragraph' ],
-				transform: ( attributes ) =>
-					createBlock( 'core/paragraph', attributes ),
-			},
-		],
-	},
-
+	transforms,
 	edit,
-
-	save( { attributes } ) {
-		const { textAlign, content } = attributes;
-
-		return (
-			<RichText.Content
-				tagName="pre"
-				style={ { textAlign } }
-				value={ content }
-			/>
-		);
-	},
-
+	save,
 	merge( attributes, attributesToMerge ) {
 		return {
 			content: attributes.content + attributesToMerge.content,
